migrate test to CompilationUnitTest
